refactor: share exact person resolution
