+++ src/domain/schemas/shared.ts
@@ -182,12 +182,11 @@ export const PersonName = NonEmptyString.pipe(Schema.brand("PersonName"))
 export type PersonName = Schema.Schema.Type<typeof PersonName>
 
 /**
- * Input schema for any field that resolves through findPersonByEmailOrName —
- * accepts either an email address or a display name. Use this for assignee,
- * lead, and similar person-reference inputs. Email validation stays strict
- * for fields where only an email makes sense (account creation, social-id
- * lookups). Named `*Input` to distinguish from the existing PersonRef
- * output struct in domain/schemas/issues.ts.
+ * Input schema for any field that accepts either an email address or a display
+ * name as a person reference. Email validation stays strict for fields where
+ * only an email makes sense (account creation, social-id lookups). Named
+ * `*Input` to distinguish from the existing PersonRef output struct in
+ * domain/schemas/issues.ts.
  */
 export const PersonRefInput = Schema.Union(Email, PersonName)
 export type PersonRefInput = Schema.Schema.Type<typeof PersonRefInput>
